+++ src/io/github/malapert/jwcs/coordsystem/FK4.java
@@ -40,22 +40,22 @@ public class FK4 implements ReferenceSystemInterface {
     /**
      * The default value of the equinox.
      */      
-    private final static float DEFAULT_EQUINOX = 1950f;
+    private final static double DEFAULT_EQUINOX = 1950d;
     
     /**
      * The default value of the epoch of observation.
      */      
-    private final static float DEFAULT_EPOCH = 1950f;
+    private final static double DEFAULT_EPOCH = 1950d;
     
     /**
      * The epoch of the equinox.
      */    
-    private float equinox;
+    private double equinox;
     
     /**
      * The epoch of observation.
      */    
-    private float epochObs;
+    private double epochObs;
     
     /**
      * Creates a FK4 reference frame with default values of both equinox
@@ -69,7 +69,7 @@ public class FK4 implements ReferenceSystemInterface {
      * Creates a FK4 reference frame with a equinox.
      * @param equinox the equinox
      */
-    public FK4(float equinox) {       
+    public FK4(double equinox) {       
         init(equinox, DEFAULT_EPOCH);
     }
     
@@ -78,7 +78,7 @@ public class FK4 implements ReferenceSystemInterface {
      * @param equinox the equinox
      * @param epoch the epoch of observation
      */
-    public FK4(float equinox, float epoch) {       
+    public FK4(double equinox, double epoch) {       
         init(equinox, epoch);
     }
     
@@ -87,7 +87,7 @@ public class FK4 implements ReferenceSystemInterface {
      * @param equinox the equinox
      * @param epochObs the epoch of observation
      */
-    private void init(float equinox, float epochObs) {
+    private void init(double equinox, double epochObs) {
         this.setEpochObs(epochObs);
         this.setEquinox(equinox);
     }
@@ -109,7 +109,7 @@ public class FK4 implements ReferenceSystemInterface {
      * @param epoch A Besselian epoch
      * @return A tuple containing the e-terms vector (DeltaD,DeltaC,DeltaC.tan(e0))
      */
-    public final static RealMatrix getEterms(float epoch) {
+    public final static RealMatrix getEterms(double epoch) {
         //Julian centuries since B1950
         double T = (epoch-1950.0d)*1.00002135903d/100.0d;
         //Eccentricity of the Earth's orbit
@@ -133,12 +133,12 @@ public class FK4 implements ReferenceSystemInterface {
     } 
 
     @Override
-    public Float getEpochObs() {
+    public Double getEpochObs() {
         return this.epochObs;
     }
 
     @Override
-    public float getEquinox() {
+    public double getEquinox() {
         return this.equinox;
     }   
 
@@ -151,7 +151,7 @@ public class FK4 implements ReferenceSystemInterface {
      * Sets the equinox.
      * @param equinox the equinox to set
      */
-    public void setEquinox(float equinox) {
+    public void setEquinox(double equinox) {
         this.equinox = equinox;
     }
 
@@ -159,7 +159,7 @@ public class FK4 implements ReferenceSystemInterface {
      * Sets the epoch.
      * @param epochObs the epochObs to set
      */
-    public void setEpochObs(float epochObs) {
+    public void setEpochObs(double epochObs) {
         this.epochObs = epochObs;
     }
