+++ src/io/github/malapert/jwcs/coordsystem/ICRS.java
@@ -46,12 +46,12 @@ public class ICRS implements ReferenceSystemInterface {
     /**
      * The default value of the equinox.
      */
-    private final static float DEFAULT_EQUINOX = 2000.0f;
+    private final static double DEFAULT_EQUINOX = 2000.0d;
     
     /**
      * The epoch of the equinox.
      */
-    private float equinox;
+    private double equinox;
 
     /**
      * Creates a new ICRS reference frame.
@@ -64,7 +64,7 @@ public class ICRS implements ReferenceSystemInterface {
      * initialization.
      * @param equinox the equinox
      */
-    private void init(float equinox) {
+    private void init(double equinox) {
         this.equinox = equinox;
     }
 
@@ -74,12 +74,12 @@ public class ICRS implements ReferenceSystemInterface {
     }
 
     @Override
-    public Float getEpochObs() {
+    public Double getEpochObs() {
         return null;
     }
 
     @Override
-    public float getEquinox() {
+    public double getEquinox() {
         return this.equinox;
     }
